Remove deprecated exports
