+++ src/__tests__/deprecated.spec.ts
@@ -4,8 +4,7 @@
  * maintained until that time.
  */
 import { describe, it, expect, beforeEach } from 'vitest';
-import * as fixtures from '../__fixtures__/fixtures.js';
-import cheerio from '../index.js';
+import { cheerio, food, fruits } from '../__fixtures__/fixtures.js';
 
 describe('deprecated APIs', () => {
   describe('cheerio module', () => {
@@ -95,7 +94,7 @@ describe('deprecated APIs', () => {
       let $: typeof cheerio;
 
       beforeEach(() => {
-        $ = cheerio.load(fixtures.food);
+        $ = cheerio.load(food);
       });
 
       it('(container, contained) : should correctly detect the provided element', () => {
@@ -131,7 +130,7 @@ describe('deprecated APIs', () => {
 
   describe('Cheerio function', () => {
     it('.load', () => {
-      const $1 = cheerio.load(fixtures.fruits);
+      const $1 = cheerio.load(fruits);
       const $2 = $1.load('<div><p>Some <a>text</a>.</p></div>');
 
       expect($2('a')).toHaveLength(1);
@@ -173,7 +172,7 @@ describe('deprecated APIs', () => {
       });
 
       it('(selector) : should return the outerHTML of the selected element', () => {
-        const $ = cheerio.load(fixtures.fruits);
+        const $ = cheerio.load(fruits);
         expect($.html('.pear')).toBe('<li class="pear">Pear</li>');
       });
     });
