+++ src/api/css.spec.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import cheerio, { load, type Cheerio } from '../index.js';
+import { load, type Cheerio } from '../index.js';
 import type { Element } from 'domhandler';
-import { mixedText } from '../__fixtures__/fixtures.js';
+import { cheerio, mixedText } from '../__fixtures__/fixtures.js';
 
 describe('$(...)', () => {
   describe('.css', () => {
