+++ src/index-browser.mts
@@ -1,3 +1,6 @@
-export * from './base-exports.js';
-// TODO: Remove this
-export { default } from './base-exports.js';
+export type { Cheerio } from './cheerio.js';
+export type * from './types.js';
+export type { CheerioOptions, HTMLParser2Options } from './options.js';
+export type { CheerioAPI } from './load.js';
+
+export * from './load-parse.js';
