+++ src/static.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import * as fixtures from './__fixtures__/fixtures.js';
-import cheerio, { type CheerioAPI } from './index.js';
+import { cheerio, food, eleven } from './__fixtures__/fixtures.js';
+import { type CheerioAPI } from './index.js';
 
 describe('cheerio', () => {
   describe('.html', () => {
@@ -270,7 +270,7 @@ describe('cheerio', () => {
     let $: CheerioAPI;
 
     beforeEach(() => {
-      $ = cheerio.load(fixtures.food);
+      $ = cheerio.load(food);
     });
 
     it('(container, contained) : should correctly detect the provided element', () => {
@@ -307,7 +307,7 @@ describe('cheerio', () => {
 
   describe('.extract', () => {
     it('() : should extract values for selectors', () => {
-      const $ = cheerio.load(fixtures.eleven);
+      const $ = cheerio.load(eleven);
 
       expect(
         $.extract({
